requirements.txt
# FactAgent - Dependencies
# AI / LLM
anthropic>=0.42.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# Web Search
tavily-python>=0.5.0

# Structured Output
pydantic>=2.0.0

# Frontend
chainlit>=1.3.0

# Utilities
python-dotenv>=1.0.0
